changed decimals to integers

+++ db/migrate/20111103005133_add_age_to_calf.rb
@@ -1,6 +1,6 @@
 class AddAgeToCalf < ActiveRecord::Migration
   def self.up
-    add_column :calves, :age, :double
+    add_column :calves, :age, :integer
   end
 
   def self.down
